Coerce to vegaspec

Vega and Vega-Lite use JSON as the format for their specifications. Within R, it seems natural to work with these specifications as lists. Accordingly, a vegaspec is also a list. This family of functions is used to coerce lists, JSON, and character strings to vegaspec.

as_vegaspec(spec, ...)

# S3 method for default
as_vegaspec(spec, ...)

# S3 method for vegaspec
as_vegaspec(spec, ...)

# S3 method for list
as_vegaspec(spec, ...)

# S3 method for json
as_vegaspec(spec, ...)

# S3 method for character
as_vegaspec(spec, ...)

Arguments

spec
object to be coerced to vegaspec, a Vega/Vega-Lite specification

...
other args (attempt to future-proof)

Value

S3 object of class vegaspec

Details

The character method for this function will take:

JSON string

A path to a local JSON file

A URL that contains a JSON file, requires httr be installed

For Vega and Vega-Lite, the translation between list and JSON is a little bit particular. This function, as_vegaspec(), can be used to translate from JSON; vw_as_json() can be used to translate to JSON.

A given vegaspec will be particular to either Vega or Vega-Lite. You can use the function vw_spec_version() to determine this. You can use vw_to_vega() to translate a Vega-Lite spec to Vega.

See also

Examples

  
# NOT RUN {
    as_vegaspec("https://vega.github.io/vega-lite/examples/specs/bar.vl.json")

# }