Create a gist that can be used as a block

Create a gist that can be used as a block

create_block(spec, embed = vega_embed(), block = block_yaml(),
  versions = vega_versions(major = FALSE), description = "",
  readme = NULL, use_thumbnail = TRUE, use_preview = TRUE,
  git_method = c("ssh", "https"), host = NULL, env_pat = NULL)

Arguments

spec
json, list, or character Vega-Lite or Vega specification

embed
vega_embed object (not yet implemented)

block
character YAML text for the .block file

versions
named list of character: names refer to JavaScript libraries vega, vega_lite, vega_embed, values are the tags for the versions to use in the block - use the helper function vega_versions() with major = TRUE to use the current major versions rather than the versions supported in this package.

description
character description for the gist

readme
character single line, path to a markdown file; multiple lines, markdown text (not yet implemented)

use_thumbnail
logical indicates to include a thumbnail image (not yet implemented)

use_preview
logical indicates to include a preview image (not yet implemented)

git_method
character use "ssh" or "https"

host
`character`` Base endpoint for GitHub API, defaults to "https://api.github.com". Useful to specify with GitHub Enterprise, e.g. "https://github.acme.com/api/v3". (not yet implemented)

env_pat
(character) Name of environment variable that contains a GitHub PAT (Personal Access Token), defaults to "GITHUB_PAT". Useful to specify with GitHub Enterprise, e.g. "GITHUB_ACME_PAT". (not yet implemented)

Value

called for side effects

See also