Create or write image

If the webdriver package and PhantomJS are installed, these functions can be used to create or write images as PNG or SVG, using a vegaspec or vegawidget.

vw_to_svg(...)

# S3 method for default
vw_to_svg(spec, scale = 1, embed = NULL,
  width = NULL, height = NULL, ...)

# S3 method for vegawidget
vw_to_svg(widget, scale = 1, ...)

vw_to_png(...)

# S3 method for default
vw_to_png(spec, scale = 1, embed = NULL,
  width = NULL, height = NULL, ...)

# S3 method for vegawidget
vw_to_png(widget, scale = 1, ...)

vw_write_svg(...)

# S3 method for default
vw_write_svg(spec, path, scale = 1, embed = NULL,
  width = NULL, height = NULL, ...)

# S3 method for vegawidget
vw_write_svg(widget, path, scale = 1, ...)

vw_write_png(...)

# S3 method for default
vw_write_png(spec, path, scale = 1, embed = NULL,
  width = NULL, height = NULL, ...)

# S3 method for vegawidget
vw_write_png(widget, path, scale = 1, ...)

Arguments

...
other arguments passed to htmlwidgets::createWidget()

spec
object to be coerced to vegaspec, a Vega/Vega-Lite specification

scale
numeric, scale-factor for the image: ratio of the width (pixels) of the image to the width (pixels) of the rendered chart

embed
list to specify vega-embed options, see Details on how this is set if NULL.

width
integer, if specified, the total rendered width (in pixels) of the chart - valid only for single-view charts and layered charts; the default is to use the width in the chart specification

height
integer, if specified, the total rendered height (in pixels) of the chart - valid only for single-view charts and layered charts; the default is to use the height in the chart specification

widget
vegawidget, created using vegawidget()

path
character, local path to which to write file

Value

vw_to_png()

character, data-URI string for PNG

vw_to_svg()

character, SVG string

vw_write_png()

invisible vegaspec or vegawidget

vw_write_svg()

invisible vegaspec or vegawidget

Details

These functions can be called using (an object that can be coerced to) a vegaspec or a vegawidget. In addition to the webdriver package, you may also need to install PhantomJS; you can use webdriver::install_phantomjs().

Images are created using the vega-view image-export functions. These functions accept a scale factor, represented here as the scale argument. By specifying scale = 2 for the PNG functions, a retina-ready image will be generated.

The vw_to_png() function returns a data-URI string for the PNG image. If you want this PNG image as raw binary data, use vw_to_png() followed by vw_png_bin().

See also

Examples

# NOT RUN {
  # call any of these functions using either a vegaspec or a vegawidget
  vw_to_svg(vegawidget(spec_mtcars))
  write_png(spec_mtcars, "temp.png")
  spec_mtcars %>% vw_to_png() %>% vw_png_bin()
# }