Create a Vega/Vega-Lite htmlwidget

The main use of this package is to render a vegawidget, which is also an htmlwidget. This function builds a vegawidget using a vegaspec.

vegawidget(spec, embed = NULL, width = NULL, height = NULL, ...)

Arguments

spec
object to be coerced to vegaspec, a Vega/Vega-Lite specification

embed
list to specify vega-embed options, see Details on how this is set if NULL.

width
integer, if specified, the total rendered width (in pixels) of the chart - valid only for single-view charts and layered charts; the default is to use the width in the chart specification

height
integer, if specified, the total rendered height (in pixels) of the chart - valid only for single-view charts and layered charts; the default is to use the height in the chart specification

...
other arguments passed to htmlwidgets::createWidget()

Value

S3 object of class vegawidget and htmlwidget

Details

If embed is NULL, vegawidget() uses:

getOption("vega.embed"), if that is NULL:

an empty call to vega_embed()

The most-important arguments to vega_embed() are:

renderer, to specify "canvas" (default) or "svg"

actions, to specify action-links for export, source, compiled, and editor

If either width or height is specified, the autosize() function is used to override the width and height of the spec. There are some important provisions:

Specifying width and height is effective only for single-view charts and layered charts. It will not work for contatenated, faceted, or repeated charts.

In the spec, the default interpretation of width and height is to describe the dimensions of the plotting rectangle, not including the space used by the axes, labels, etc. Here, width and height describe the dimensions of the entire rendered chart, including axes, labels, etc.

See also

Examples

  vegawidget(spec_mtcars, width = 350, height = 350)