Shiny-output for vegawidget

Use this function in the UI part of your Shiny app.

vegawidgetOutput(outputId, width = "auto", height = "auto", widget = NULL)

Arguments

outputId

output variable to read from

width, height

Must be a valid CSS unit (like "100%", "400px", "auto") or a number, which will be coerced to a string and have "px" appended. For vegawidgets, "auto" is useful because, as of now, the spec determines the size of the widget, then the widget determines the size of the container.

widget

character, indicating which version of libraries to use, e.g. "vl5". Normally, you should not need to set this. See vega_version_all() for more information.